pyproject.toml
[project]
name = "my_project"
version = "0.1.0"
dependencies = [
    "boto3>=1.36.6",
    "fastapi>=0.115.7",
    "matplotlib>=3.10.0",
    "mlflow>=2.20.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-binance>=1.0.27",
    "requests",
    "scikit-learn>=1.6.1",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "uvicorn>=0.34.0",
]
